requirements.txt
fastapi
uvicorn
pydantic
requests
google-generativeai

pytest